+++ src/datumaro/components/dataset_item_storage.py
@@ -1,0 +1,160 @@
+from __future__ import annotations
+
+from copy import copy
+from enum import Enum, auto
+from typing import Iterator, Union, Optional, Any, Tuple, Type
+
+from datumaro.components.dataset_base import DatasetItem, DEFAULT_SUBSET_NAME, IDataset, CategoriesInfo, MediaElement
+
+__all__ = ["ItemStatus", "DatasetItemStorage", "DatasetItemStorageDatasetView"]
+
+
+class ItemStatus(Enum):
+    added = auto()
+    modified = auto()
+    removed = auto()
+
+
+class DatasetItemStorage:
+    def __init__(self):
+        self.data = {}  # { subset_name: { id: DatasetItem } }
+        self._traversal_order = {}  # maintain the order of elements
+
+    def __iter__(self) -> Iterator[DatasetItem]:
+        for item in self._traversal_order.values():
+            yield item
+
+    def __len__(self) -> int:
+        return len(self._traversal_order)
+
+    def is_empty(self) -> bool:
+        # Subsets might contain removed items, so this may differ from __len__
+        return all(len(s) == 0 for s in self.data.values())
+
+    def put(self, item: DatasetItem) -> bool:
+        subset = self.data.setdefault(item.subset, {})
+        is_new = subset.get(item.id) is None
+        self._traversal_order[(item.id, item.subset)] = item
+        subset[item.id] = item
+        return is_new
+
+    def get(
+        self, id: Union[str, DatasetItem], subset: Optional[str] = None, dummy: Any = None
+    ) -> Optional[DatasetItem]:
+        if isinstance(id, DatasetItem):
+            id, subset = id.id, id.subset
+        else:
+            id = str(id)
+            subset = subset or DEFAULT_SUBSET_NAME
+
+        return self.data.get(subset, {}).get(id, dummy)
+
+    def remove(self, id: Union[str, DatasetItem], subset: Optional[str] = None) -> bool:
+        if isinstance(id, DatasetItem):
+            id, subset = id.id, id.subset
+        else:
+            id = str(id)
+            subset = subset or DEFAULT_SUBSET_NAME
+
+        subset_data = self.data.setdefault(subset, {})
+        is_removed = subset_data.get(id) is not None
+        subset_data[id] = None
+        if is_removed:
+            self._traversal_order.pop((id, subset))
+        return is_removed
+
+    def __contains__(self, x: Union[DatasetItem, Tuple[str, str]]) -> bool:
+        if not isinstance(x, tuple):
+            x = [x]
+        dummy = 0
+        return self.get(*x, dummy=dummy) is not dummy
+
+    def get_subset(self, name):
+        return self.data.get(name, {})
+
+    def subsets(self):
+        return self.data
+
+    def __copy__(self):
+        copied = DatasetItemStorage()
+        copied._traversal_order = copy(self._traversal_order)
+        copied.data = copy(self.data)
+        return copied
+
+
+class DatasetItemStorageDatasetView(IDataset):
+    class Subset(IDataset):
+        def __init__(self, parent: DatasetItemStorageDatasetView, name: str):
+            super().__init__()
+            self.parent = parent
+            self.name = name
+
+        @property
+        def _data(self):
+            return self.parent._get_subset_data(self.name)
+
+        def __iter__(self):
+            for item in self._data.values():
+                if item:
+                    yield item
+
+        def __len__(self):
+            return len(self._data)
+
+        def put(self, item):
+            return self._data.put(item)
+
+        def get(self, id, subset=None):
+            assert (subset or DEFAULT_SUBSET_NAME) == (self.name or DEFAULT_SUBSET_NAME)
+            return self._data.get(id, subset)
+
+        def remove(self, id, subset=None):
+            assert (subset or DEFAULT_SUBSET_NAME) == (self.name or DEFAULT_SUBSET_NAME)
+            return self._data.remove(id, subset)
+
+        def get_subset(self, name):
+            assert (name or DEFAULT_SUBSET_NAME) == (self.name or DEFAULT_SUBSET_NAME)
+            return self
+
+        def subsets(self):
+            return {self.name or DEFAULT_SUBSET_NAME: self}
+
+        def categories(self):
+            return self.parent.categories()
+
+        def media_type(self):
+            return self.parent.media_type()
+
+    def __init__(
+        self,
+        parent: DatasetItemStorage,
+        categories: CategoriesInfo,
+        media_type: Optional[Type[MediaElement]],
+    ):
+        self._parent = parent
+        self._categories = categories
+        self._media_type = media_type
+
+    def __iter__(self):
+        yield from self._parent
+
+    def __len__(self):
+        return len(self._parent)
+
+    def categories(self):
+        return self._categories
+
+    def get_subset(self, name):
+        return self.Subset(self, name)
+
+    def _get_subset_data(self, name):
+        return self._parent.get_subset(name)
+
+    def subsets(self):
+        return {k: self.get_subset(k) for k in self._parent.subsets()}
+
+    def get(self, id, subset=None):
+        return self._parent.get(id, subset=subset)
+
+    def media_type(self):
+        return self._media_type
