+++ src/datumaro/components/operations.py
@@ -41,7 +41,8 @@ from datumaro.components.annotation import (
     PointsCategories,
 )
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset import Dataset, DatasetItemStorage, IDataset
+from datumaro.components.dataset import Dataset, IDataset
+from datumaro.components.dataset_item_storage import DatasetItemStorage
 from datumaro.components.dataset_base import CategoriesInfo, DatasetItem
 from datumaro.components.errors import (
     AnnotationsTooCloseError,
