+++ src/datumaro/plugins/data_formats/cifar.py
@@ -10,7 +10,7 @@ from collections import OrderedDict
 import numpy as np
 
 from datumaro.components.annotation import AnnotationType, Label, LabelCategories
-from datumaro.components.dataset import ItemStatus
+from datumaro.components.dataset_item_storage import ItemStatus
 from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
 from datumaro.components.errors import MediaTypeError
 from datumaro.components.exporter import Exporter
