+++ src/datumaro/plugins/data_formats/open_images.py
@@ -20,7 +20,7 @@ import numpy as np
 from attr import attrs
 
 from datumaro.components.annotation import AnnotationType, Bbox, Label, LabelCategories, Mask
-from datumaro.components.dataset import ItemStatus
+from datumaro.components.dataset_item_storage import ItemStatus
 from datumaro.components.dataset_base import DatasetBase, DatasetItem, Importer
 from datumaro.components.errors import (
     DatasetError,
