+++ src/datumaro/plugins/data_formats/sly_pointcloud/exporter.py
@@ -15,7 +15,7 @@ import uuid
 from datetime import datetime
 
 from datumaro.components.annotation import AnnotationType, LabelCategories
-from datumaro.components.dataset import ItemStatus
+from datumaro.components.dataset_item_storage import ItemStatus
 from datumaro.components.dataset_base import DatasetItem, IDataset
 from datumaro.components.errors import MediaTypeError
 from datumaro.components.exporter import Exporter
