+++ tests/unit/test_dataset.py
@@ -18,7 +18,8 @@ from datumaro.components.annotation import (
     Polygon,
     PolyLine,
 )
-from datumaro.components.dataset import DEFAULT_FORMAT, Dataset, ItemStatus, eager_mode
+from datumaro.components.dataset import DEFAULT_FORMAT, Dataset, eager_mode
+from datumaro.components.dataset_item_storage import ItemStatus
 from datumaro.components.dataset_base import (
     DEFAULT_SUBSET_NAME,
     DatasetBase,
